requirements.txt
Flask>=2.2
Werkzeug>=2.2.0
Flask-SQLAlchemy==3.0
psycopg2-binary==2.9.1